requirements.txt
# openai==1.3.4
# openai
# tiktoken
streamlit==1.31.0
# Extra-Streamlit-Components
# streamlit-qrcode-scanner
# streamlit-option-menu
# st-pages
python-dotenv
# numpy
pydantic
requests
# python-docx
# AudioSegment
# pydub
# soundfile 
loguru
# pyjwt
# streamlit-mic-recorder
# pydub
# PyYAML
# sqlalchemy
# beanie
# motor
# pymongo